(INclude new sql!!!) Fix search filter, prepare all sql statements, salt and hash password.  etc. etc.

+++ php/updateGuide.php
@@ -10,6 +10,8 @@
 
 <?php
 	
+	include 'createGuide.php';
+
 	$mysqli = new mysqli('localhost', 'root', '', 'diy_tour');
 	
 	if ($mysqli->connect_error) {
@@ -17,19 +19,20 @@
 	}
 		
 	
-	if(isset($_GET['numDay'])&& isset($_GET['id'])){
-		updateGuide ($_GET['numDay'],$_GET['id'], $mysqli);
+	if(isset($_POST['numDay'])&& isset($_POST['id'])){
+		updateGuide ($_POST['numDay'],$_POST['id'], $mysqli);
 	}
 	
 	
 	function changeFeatureImage($image,$mysqli,$guideId){
-		$target_dir = "../guide_Image/";
-		$target_file = $target_dir . basename($image["name"]);
-		$imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
-		
+
 		if(isset($_POST['feaureImageDelete'])&& $_POST['feaureImageDelete']!=null){
-			$query = mysqli_query($mysqli,"UPDATE travelGuide SET featureImage = 'guide_Image/NoPicAvailable.png'
-			WHERE guideId =".$guideId);
+			$sql = "UPDATE travelGuide SET featureImage = 'guide_Image/NoPicAvailable.png'
+			WHERE guideId = ?";
+			$stmt = $mysqli->prepare($sql);
+			$stmt->bind_param("i", $guideId);
+			$stmt->execute();
+			
 			$image = "../". (string)$_POST['feaureImageDelete'];
 			if (file_exists($image)) 
                {
@@ -41,47 +44,31 @@
                echo "File does not exists"; 
               }
 		}else if ($_FILES['main-upload']['size'] != 0){
+			$imageFileType = getImageType("../guide_Image/",$image["name"]);
 			$target_loc = "guide_Image/"."GuideID".$guideId.".".$imageFileType;
-			$query = mysqli_query($mysqli,"UPDATE travelGuide SET featureImage = '$target_loc'
-			WHERE guideId =".$guideId);
 			
-			if ($query){
-				if (move_uploaded_file($image["tmp_name"], "../".$target_loc)) {
-					echo $target_loc;
-				} else {
-					echo "Sorry, there was an error uploading your file.";
-				}
-			}
+			$sql = "UPDATE travelGuide SET featureImage = ? WHERE guideId = ?";
+			$stmt = $mysqli->prepare($sql);
+			$stmt->bind_param("si", $target_loc,$guideId);
+			
+			uploadImage($target_loc,$image["tmp_name"],$stmt);
 		}
 	}
 	
-	function addImage($image,$id,$count,$increase,$mysqli){
-		$target_dir = "../guide_Image/";
-		$target_file = $target_dir . basename($image["name"][$count]);
-		echo $image["name"][$count];
-		$imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
-	
-		$target_loc = "guide_Image/"."DayID".$id."_".($count+$increase).".".$imageFileType;
-		$query = mysqli_query($mysqli,"INSERT INTO image (imageLink,dayId) VALUES ('$target_loc','$id')");
-		
-		if ($query){
-			if (move_uploaded_file($image["tmp_name"][$count], "../".$target_loc)) {
-				return true;
-			} else {
-				echo "Sorry, there was an error uploading your file.";
-				return false;
-			}
-		}
-	}
 	
 	function deleteImage($guideId, $mysqli){
 		$num = 0;
-		while (isset($_POST['image'.$num])&& $_POST['image'.$num]!="" && $_POST['image'.$num]!=null){
+		while (isset($_POST['image'.$num]) && $_POST['image'.$num]!="" && $_POST['image'.$num]!=null){
 			echo($_POST['image'.$num]);
-			$query = mysqli_query($mysqli,"DELETE FROM image WHERE imageLink ='".$_POST['image'.$num]."'");
-			echo "DELETE FROM image WHERE imageLink =".$_POST['image'.$num];
-			$image = "../".(string)$_POST['image'.$num];
-			unlink($image);
+
+			$sql = "DELETE FROM image WHERE imageLink = ?";
+			$stmt = $mysqli->prepare($sql);
+			$stmt->bind_param("s", $_POST['image'.$num]);
+			
+			if ($stmt->execute()){
+				$image = "../".(string)$_POST['image'.$num];
+				unlink($image);
+			}
 			$num++;
 		}
 	}
@@ -94,50 +81,70 @@
 		$budget= $_POST['budget'];
 		$summary= $_POST['summary'];
 		
-		session_start();
 		$id = $_SESSION['id'];
 		
-		$query = mysqli_query($mysqli,"UPDATE travelguide SET guideName = '$title'
-		,country = '$loc', date = '$date', people ='$people'
-		,budget = '$budget', overview = '$summary'
-		WHERE guideId = '$guideId' AND userId = '$id'");
+		$sql = "UPDATE travelguide SET guideName = ?,country = ?, date = ?, people = ?
+		,budget = ?, overview = ? WHERE guideId = ? AND userId = ?";
+		$stmt = $mysqli->prepare($sql);
+		$stmt->bind_param("ssssisii", $title,$loc,$date,$people,$budget,$summary,$guideId,$id );
 		
-		$featureimage =  $_FILES['main-upload'];
-		changeFeatureImage($featureimage,$mysqli,$guideId);
+		if ($stmt->execute()){
+			$featureimage =  $_FILES['main-upload'];
+			changeFeatureImage($featureimage,$mysqli,$guideId);
 
-		deleteImage($guideId, $mysqli);
-		
+			deleteImage($guideId, $mysqli);
+			
+			updateDay($num,$guideId,$title,$mysqli);
+		}
+	}
+
+	function updateDay($num,$guideId,$title,$mysqli){
 		for ($i = 1; $i <= $num; $i++) {
 			$dtitle = $_POST['title'.$i];
 			$dinfo =  $_POST['summary'.$i];
 			
-			$sql = "UPDATE day SET Title = '$dtitle'
-			,description = '$dinfo' WHERE dayNum = '$i' AND guideId = '$guideId'";
-			$querySelect = mysqli_query($mysqli,"SELECT dayId FROM day WHERE dayNum = '$i' AND guideId = '$guideId'");
-			$dayId=mysqli_fetch_row($querySelect)[0];
-echo mysqli_num_rows($querySelect);
-			if (mysqli_num_rows($querySelect)<1) {
+
+			$sql = "UPDATE day SET Title = ?,description = ? WHERE dayNum = ? AND guideId = ?";
+			
+			$stmt = $mysqli->prepare($sql);
+			$stmt->bind_param("ssii", $dtitle,$dinfo,$i,$guideId);
+			
+			$daySql = "SELECT dayId FROM day WHERE dayNum = ? AND guideId = ?";
+			$dayStmt = $mysqli->prepare($daySql);
+			$dayStmt->bind_param("ii", $i,$guideId);
+			$dayStmt->execute();
+			$result = $dayStmt->get_result();
+			$dayId=mysqli_fetch_row($result)[0];
+			
+			if (mysqli_num_rows($result)<1) {
 				echo "add new";
-				$sqlInsert = "INSERT INTO day (guideId, title, description,dayNum) VALUES ('$guideId','$dtitle','$dinfo','$i')";
-				$dayId = mysqli_insert_id($mysqli);
-				if ($mysqli->query($sqlInsert) === FALSE) {
+				$newStmt = createDay($dtitle,$dinfo,$guideId,$mysqli,$i);
+				
+				if ($newStmt->execute()) {
 					echo "Error update your guide. Please try again later.";
 					break;
 				}
+				$dayId = mysqli_insert_id($mysqli);
 			}
 			
-			if ($mysqli->query($sql) === TRUE) {
+			if ($stmt->execute()) {
 				$countfiles = count((array_filter($_FILES['image-upload'.$i]['name']))); 
-				$result=mysqli_query($mysqli,"SELECT count(*) as total from image WHERE dayId = ".$dayId);
-				$data=mysqli_fetch_assoc($result);
-				echo "_????_";
+
+				$countSql = "SELECT count(*) as total FROM image WHERE dayId = ?";
+				$countStmt = $mysqli->prepare($countSql);
+				$countStmt->bind_param("i", $dayId);
+				$countStmt->execute();
+				$countResult = $countStmt->get_result();
+
+				$data=mysqli_fetch_assoc($countResult);
 				echo $i;
-				echo "_????_";
 				for($j=0; $j<$countfiles; $j++){
 					$image =  $_FILES['image-upload'.$i];
 					
-					if (!addImage ($image,$dayId,$j,$data['total']+1,$mysqli)){
-						return;
+					if (isset($image) && $image!="" && $image!=null){
+						if (!appendImage ($image,$dayId,$j,$data['total']+1,$mysqli)){
+							return;
+						}
 					}
 				}
 			}else{
@@ -146,19 +153,31 @@ echo mysqli_num_rows($querySelect);
 			}
 			
 		}
-		deleteDay ($i,$guideId,$mysqli);
+		deleteDay ($num,$guideId,$mysqli);
 		header('Location: ../specificGuide.html?id='.$guideId.'&title='.$title);
 	}
 	
+	function appendImage($image,$id,$count,$increase,$mysqli){
+		$imageFileType = getImageType("../guide_Image/",$image["name"][$count]);
+
+		$target_loc = "guide_Image/"."DayID".$id."_".($count+$increase).".".$imageFileType;
+	
+		$sql = "INSERT INTO image (imageLink,dayId) VALUES (?,?)";
+		$stmt = $mysqli->prepare($sql);
+		$stmt->bind_param("si", $target_loc, $id);
+		
+		return uploadImage($target_loc,$image["tmp_name"][$count],$stmt);
+	}
+	
 	function deleteDay ($dayNum,$id,$mysqli){
-		$query = mysqli_query($mysqli,"DELETE FROM day WHERE guideId = '$id' AND dayNum >'$dayNum'");
-		echo "DELETE FROM day WHERE guideId = '$id' AND dayNum >'$dayNum'";
-		if (!$query){
+		$sql = "DELETE FROM day WHERE guideId = ? AND dayNum > ?";
+		$stmt = $mysqli->prepare($sql);
+		$stmt->bind_param("ii", $id,$dayNum);
+		if (!$stmt->execute()){
 			echo "Error creating new guide. Please try again later.?";
 		}
 		
 	}
-	
 	
 	$mysqli->close();
 ?>
